:bug: fix assert_str bug and pull out example code into examples directory
